refactor: Split the CLI menus into different files to make it easier to manage
