+++ src/causal-tree-cli/migrate/index.ts
@@ -1,0 +1,1 @@
+export * from './migrate';
