+++ src/causal-tree-cli/migrate/mongodb.ts
@@ -1,0 +1,43 @@
+import { prompt } from 'inquirer';
+import { MongoDBRepoStore } from '@casual-simulation/causal-tree-store-mongodb';
+import { connect } from 'mongodb';
+
+export async function mongoConnectionInfo(): Promise<MongoDBRepoStore> {
+    const urlAnswer = await prompt({
+        type: 'input',
+        name: 'url',
+        message: 'Enter the MongoDB URL',
+    });
+
+    const client = await connect(
+        urlAnswer.url,
+        { useNewUrlParser: true }
+    );
+
+    const databasesResult = (await client
+        .db()
+        .admin()
+        .listDatabases()) as {
+        databases: {
+            name: string;
+        }[];
+    };
+
+    const databases = databasesResult.databases.map(d => d.name);
+
+    const databaseAnswer = await prompt({
+        type: 'list',
+        name: 'database',
+        message: 'Select the database',
+        choices: databases,
+    });
+
+    const db = client.db(databaseAnswer.database);
+
+    const objectsCollection = db.collection('objects');
+    const headsCollection = db.collection('heads');
+
+    const mongoStore = new MongoDBRepoStore(objectsCollection, headsCollection);
+    await mongoStore.init();
+    return mongoStore;
+}
